+++ src/main/java/techcourse/myblog/service/dto/UserDto.java
@@ -1,8 +1,22 @@
-package techcourse.myblog.dto;
+package techcourse.myblog.service.dto;
 
-public class UserLoginDto {
+import techcourse.myblog.support.config.validator.EmailConstraint;
+
+public class UserDto {
+    private String name;
+
+    @EmailConstraint
     private String email;
+
     private String password;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
 
     public String getEmail() {
         return email;
