+++ src/main/java/techcourse/myblog/service/dto/UserResponseDto.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.dto;
+package techcourse.myblog.service.dto;
 
 public class UserResponseDto {
     private String name;
