+++ src/main/java/techcourse/myblog/service/exception/EditException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.exception;
+package techcourse.myblog.service.exception;
 
 public class EditException extends RuntimeException {
     public EditException(String message) {
