+++ src/main/java/techcourse/myblog/service/exception/ErrorMessage.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.exception;
+package techcourse.myblog.service.exception;
 
 public class ErrorMessage {
     private String errorMessage;
