+++ src/main/java/techcourse/myblog/service/exception/LoginException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.exception;
+package techcourse.myblog.service.exception;
 
 public class LoginException extends RuntimeException {
     public LoginException(String message) {
