+++ src/main/java/techcourse/myblog/service/exception/SignUpException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.exception;
+package techcourse.myblog.service.exception;
 
 public class SignUpException extends RuntimeException {
     public SignUpException(String message) {
