+++ src/main/java/techcourse/myblog/support/config/WebConfig.java
@@ -1,9 +1,9 @@
-package techcourse.myblog.config;
+package techcourse.myblog.support.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import techcourse.myblog.interceptor.AuthInterceptor;
+import techcourse.myblog.web.interceptor.AuthInterceptor;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
